backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
pytest==8.3.5
httpx==0.28.1
python-dotenv==1.1.0
